refactor: consolidate type imports to reduce type issues across the codebase
