+++ src/hooks/actions/use-msg-create-appointment.tsx
@@ -16,7 +16,7 @@ import { isDraft, isSpam } from 'helpers/folders';
 import { useUiUtilities } from 'hooks/use-ui-utilities';
 import { extractBody } from 'store/editor-slice-utils';
 import { getMessageEmailStoreAction } from 'store/emails/actions/get-message';
-import { CalendarType, SenderType } from 'types/calendar/index.d';
+import { CalendarType, SenderType } from 'types/calendar';
 import type { ActionFn, MailMessage, UIActionDescriptor } from 'types/index.d';
 
 export const useMsgCreateAppointmentFn = (item: MailMessage, folderId: string): ActionFn => {
