+++ src/types/state/index.d.ts
@@ -5,8 +5,8 @@
  */
 
 import { type API_REQUEST_STATUS } from 'constants/index';
-import type { SavedAttachment, UnsavedAttachment } from 'types/attachments/index.d';
-import type { Conversation } from 'types/conversations/index.d';
+import type { SavedAttachment, UnsavedAttachment } from 'types/attachments';
+import type { Conversation } from 'types/conversations';
 import { type AttachmentUploadProcessStatus, type MailsEditorV2 } from 'types/editor/index.d';
 import { type MailMessage } from 'types/messages/index.d';
