+++ src/views/settings/filters/parts/filter-actions/action-mark-as-component.tsx
@@ -7,8 +7,8 @@
 import React, { useCallback, useMemo } from 'react';
 
 import { useTranslation } from 'react-i18next';
+import { FilterFlag, MarkAsOption } from 'types/filters';
 
-import { FilterFlag, MarkAsOption } from 'types/index.d';
 import { MarkAs } from 'views/settings/filters/parts/filter-actions/mark-as';
 import { getMarkAsOptions } from 'views/settings/filters/parts/filter-actions/mark-as-utils';
 import { ActionComponentProps } from 'views/settings/filters/types';
